Fix coordinate and index arithmetic overflow on extreme inputs
line(), circle(), filled_circle(), ring_annulus(), rect(), rect_perimeter(),
chebyshev_ring(), diamond(), and the cone routine computed point offsets
and capacity hints in i32/u32, which panic in debug (and wrap in release)
for legal-but-extreme inputs: endpoints |x1 - x| > i32::MAX, centres near
i32::MIN/MAX, perimeters of huge rects. Widen accumulators to i64/usize
and saturate the emitted point coordinates. terminal.rs and tilemap.rs
index/capacity math gets the same usize treatment, matching the idiom
already used elsewhere in the crate.

+++ izanagi_kit/src/terminal.rs
@@ -99,7 +99,7 @@ impl Screen {
         if x < 0 || y < 0 || x as u32 >= self.width || y as u32 >= self.height {
             None
         } else {
-            Some((y as u32 * self.width + x as u32) as usize)
+            Some(y as usize * self.width as usize + x as usize)
         }
     }
 
@@ -272,7 +272,7 @@ impl Screen {
         for y in 0..self.height {
             let mut last: Option<(Color, Color)> = None;
             for x in 0..self.width {
-                let cell = &self.cells[(y * self.width + x) as usize];
+                let cell = &self.cells[y as usize * self.width as usize + x as usize];
                 if last != Some((cell.fg, cell.bg)) {
                     out.push_str(&format!(
                         "\x1b[38;2;{};{};{};48;2;{};{};{}m",
